Rough Draft Upload
This is a rough draft of my current ideas for this mod, been working on some XMLs and art for a few days now taking things from base game and adding on or altering some and this is what I got so far.  This will need custom classes however from DLLs.